// src/core/types.ts
export enum EntityType {
  /** The player-controlled Pacman entity. */
  Pacman = 'Pacman',
  /** An AI-controlled Ghost entity. */
  Ghost = 'Ghost',
}